refactor: split config modules
