+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.3] - 2025-09-29
+
+### Changed
+
+- Extracted configuration appearance, module layout, validation, and serde helper
+  logic into dedicated submodules, leaving `config.rs` as the facade while
+  preserving existing APIs.
+
+### Added
+
+- Unit tests covering appearance defaults, serde helpers, module layout
+  deserialization, and configuration validation to guard the new structure.
+
 ## [0.6.2] - 2025-09-28
 
 ### Changed
